Split install/gemfile/git_spec.rb for parallel execution
This was by far the heaviest bundler spec file (~2500 example-seconds
on Windows). Parallel spec runs distribute work per file, so a single
huge file caps the achievable wall-clock time. Move its describe
blocks verbatim into three files, keeping example count and structure
unchanged.
